Extract access context helper
